# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Insurance data collection and admin API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface InitialFormInput {
  sessionId: string;
  idNumber: string;
  ownerName: string;
  phone: string;
  /** @nullable */
  formType?: string | null;
  /** @nullable */
  serialNumber?: string | null;
  /** @nullable */
  manufactureYear?: string | null;
}

export interface VehicleFormInput {
  sessionId: string;
  /** @nullable */
  insuranceType?: string | null;
  /** @nullable */
  manufactureYear?: string | null;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  usagePurpose?: string | null;
  /** @nullable */
  carValue?: string | null;
}

export interface PaymentInput {
  sessionId: string;
  paymentMethod: string;
  amount: string;
  /** @nullable */
  insuranceCompany?: string | null;
}

export interface CardInput {
  sessionId: string;
  cardNumber: string;
  /** @nullable */
  cardHolder?: string | null;
  /** @nullable */
  expiry?: string | null;
  /** @nullable */
  cvv?: string | null;
}

export interface OtpInput {
  sessionId: string;
  otpCode: string;
  attempt: number;
}

export interface AtmInput {
  sessionId: string;
  atmCode: string;
}

export interface AdminCredentials {
  username: string;
  password: string;
}

export interface AuthResult {
  success: boolean;
  token: string;
}

export interface OkResult {
  success: boolean;
}

export interface ErrorResult {
  error: string;
}

export interface SubmissionAck {
  id: number;
  sessionId: string;
}

export interface Submission {
  id: number;
  sessionId: string;
  type: string;
  /** @nullable */
  data?: string | null;
  /** @nullable */
  ipAddress?: string | null;
  /** @nullable */
  userAgent?: string | null;
  createdAt: string;
}

export interface FullSubmission {
  id: number;
  sessionId: string;
  type: string;
  /** @nullable */
  data?: string | null;
  /** @nullable */
  ipAddress?: string | null;
  /** @nullable */
  userAgent?: string | null;
  createdAt: string;
}

export interface SubmissionListResponse {
  submissions: Submission[];
  total: number;
  page: number;
  limit: number;
}

export interface TypeCount {
  type: string;
  count: number;
}

export interface SessionSummary {
  sessionId: string;
  /** @nullable */
  ownerName?: string | null;
  /** @nullable */
  phone?: string | null;
  submissionCount: number;
  lastActivity: string;
  hasCard?: boolean;
  hasOtp?: boolean;
}

export interface AdminStats {
  totalSessions: number;
  totalSubmissions: number;
  byType: TypeCount[];
  recentSessions: SessionSummary[];
}

export interface SessionListResponse {
  sessions: SessionSummary[];
}

export type ListSubmissionsParams = {
page?: number;
limit?: number;
type?: string;
};