# lib/api-zod/src/generated/types/authResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Insurance data collection and admin API
 * OpenAPI spec version: 0.1.0
 */

export interface AuthResult {
  success: boolean;
  token: string;
}