# lib/api-zod/src/generated/types/otpInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Insurance data collection and admin API
 * OpenAPI spec version: 0.1.0
 */

export interface OtpInput {
  sessionId: string;
  otpCode: string;
  attempt: number;
}