# lib/api-zod/src/generated/types/sessionListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Insurance data collection and admin API
 * OpenAPI spec version: 0.1.0
 */
import type { SessionSummary } from './sessionSummary';

export interface SessionListResponse {
  sessions: SessionSummary[];
}