# lib/api-zod/src/generated/types/submission.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Insurance data collection and admin API
 * OpenAPI spec version: 0.1.0
 */

export interface Submission {
  id: number;
  sessionId: string;
  type: string;
  /** @nullable */
  data?: string | null;
  /** @nullable */
  ipAddress?: string | null;
  /** @nullable */
  userAgent?: string | null;
  createdAt: string;
}