# lib/api-zod/src/generated/types/vehicleFormInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Insurance data collection and admin API
 * OpenAPI spec version: 0.1.0
 */

export interface VehicleFormInput {
  sessionId: string;
  /** @nullable */
  insuranceType?: string | null;
  /** @nullable */
  manufactureYear?: string | null;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  usagePurpose?: string | null;
  /** @nullable */
  carValue?: string | null;
}